docs/PadroesDeProjeto/PadraoEmergente/requirements.txt

coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
Django==3.0.14
django-common-helpers==0.9.2
django-cors-headers==3.2.1
django-cron==0.5.1
django-environ==0.4.5
django-material==1.6.7
django-material-admin==1.6.44
django-polymorphic==2.1.2
django-rest-passwordreset==1.1.0
djangorestframework==3.11.2
fcm-django==0.3.11
psycopg2==2.9.6
psycopg2-binary==2.8.6
py==1.11.0
pycodestyle==2.6.0
